feat: 支持 TimeWindow 泛型，兼容多种整数类型
将时间窗口函数重构为泛型，提升对不同整数类型的兼容性，增强代码灵活性与复用性。

+++ types/time_test.go
@@ -14,7 +14,7 @@ func TestTimeWindow(t *testing.T) {
 	)
 
 	for i := int64(0); i < times; i++ {
-		c1 := 100000 + i
+		c1 := int64(100000 + i)
 		w1 := TimeWindow(c1, tolerance)
 		for j := int64(0); j < times/2; j++ {
 			c2 := c1 - times/4 + j
